Renamed class config to SiteConfig and refactored.
Many of the helper subroutines and data objects in the main
routine in podman_hpc.py were incorporated as attributes
into the config class, so that it could be easily passed
around as a stateful data object.  The class was renamed
SiteConfig to conform with capitalization best practices
(and just be a bit more descriptive).  This was all broken
out into a separate submodule called siteconfig.py