Reconfig podman_hpc to use Click and ext SiteConfig.
The main routine in podman_hpc.py is now using the
external SiteConfig submodule, so these definitions are removed.

Additionally, Click extension classes in click_passthrough.py
are now being used.  The main differences are that subcommands
are explicitly broken out as separate function definitions,
and there is some additional configuration for help message formatting.

Note that hook flags parsed from the site configuration modules are
dynamically added on subcommands, so they can be viewed in the
help pages.