Work around container removed error
There was an annoying race condition using the previous approach
for the shared run cleanup.

We believe the issue is the exec would take to long to exit out and
the container would get removed before it finished.  This would lead
to an annoying error and a non-zero exit.

The new method has the wrapper send a message via a socket to a
monitoring thread that handles the final cleanup.

We also brought back the containers.conf because the latest OS
update introduced some issues that required a config.