Improved configuration management
This refactors the configuration management.  It supports
setting parameters via defaults, a configuraiton file and
via environment variables.

The parameters are accessed by class attributes.  Some of
the parameters aren't yet utilized.  Those will be used in
future updates.

In addition to the code changes, there is a minimal unit test
and documentation.