Convert to using command-line approach
This converts to using the command-line to pass all of
the parameters and avoids encoding these in podman
configuration files.

There are two sets of configured arguments (default_args and
default_run_args).  These now contain all of the parameters
to do things like set the graph root, run root, additional
stores, etc.

This also contains a number of fixes to fix the unit test
and some flake8 fixes.