Fix PMI2 handling
PMI2 needs to pass through an open file descriptor.
There is a way to do this with podman but the file descriptors
need to be consecutive.  This fix will dup the file descriptor
to fd 3 (the first one after the defaults of stdin, out, and err).
It then set PMI_FD to point to the dupped fd (3).