Fix for issue121
The method used to find the required layers is incorrect.  It uses
a list of layers in the manifest which are the diff hashes.  It winds up
layers can share these.  This would lead to the wrong layer being pulled
in.  This patch uses the top layer in the image spec and then walks up
the parent tree to get the other layers.